Support development against Github autonity.py
Currently, autonity.py is always pulled in from PyPI. But for creating
development builds, it may be necessary to create builds that install
autonity.py from the Github repo instead.

This change adds a new `github` Hatch environment for developing and
building packages that reference a version of autonity.py located at the
Git remote hosted on Github and resolving using a specific git commit,
which may be a commit hash or the name of a branch.

By default, the `github` environment will track the `develop` branch of
autonity.py. This commit can be overridden with the environment variable
`AUTPY_GIT_COMMIT` _before_ the environment is first created. If an
environment already exists, and a different commit is required, the
environment must be removed and re-created.

The `build` script is added to the `github` environment for convenience.
But note that, as with environment creation, `AUTPY_GIT_COMMIT` must be
set to the desired commit _before_ calling this script.