Support development against local autonity.py
Previously, developing against a local checkout of autonity.py required
modifying the main project dependencies in `pyproject.toml`.

This change adds a new `local` Hatch environment for developing and
testing packages that reference a local git clone of autonity.py. By
default, the environment will look for the repo in the root directory
path `autonity.py`, which can be a symlink.

To use a different path, set the environment variable `AUTPY_LOCAL_PATH`
when creating the environment. If the environment exists and a new path
is desired, then the environment must be removed and re-created.