Move `protocol gov` command to top-level
The gov is a subcommand, unlike the other `protocol` commands. So in the
help test it does not stand out as a different category.

There is no out of the box way with Click to put subcommands in
different sections.

It makes sense to move the gov command to the top-level. At the
top-level using the full name `governance` seems appropriate.