Rebrand project from "aut" to "autonity-cli"
The project name "aut" can be improved. Feedback received is that it
is ambiguous when used in conversation, and often needs to be qualified
further with the redundant "tool" suffix.

This change is simultaneous with repo URL change and shortly publishing
the tool as a PyPI package.

Also, expand introduction in the README and use the term "command-line"
consistently.